+++ src/Behat/Testwork/Argument/MixedArgumentOrganiser.php
@@ -10,10 +10,8 @@
 
 namespace Behat\Testwork\Argument;
 
-use Behat\Testwork\Argument\Exception\UnknownParameterValueException;
 use ReflectionFunctionAbstract;
 use ReflectionClass;
-use ReflectionMethod;
 use ReflectionParameter;
 
 /**
@@ -35,12 +33,7 @@ final class MixedArgumentOrganiser implements ArgumentOrganiser
      */
     public function organiseArguments(ReflectionFunctionAbstract $function, array $arguments)
     {
-        $parameters = $function->getParameters();
-        $arguments = $this->prepareArguments($parameters, $arguments);
-
-        $this->validateArguments($function, $parameters, $arguments);
-
-        return $arguments;
+        return $this->prepareArguments($function->getParameters(), $arguments);
     }
 
     /**
@@ -400,55 +393,6 @@ final class MixedArgumentOrganiser implements ArgumentOrganiser
         }
 
         return $orderedArguments;
-    }
-
-    /**
-     * Validates that all arguments are in place, throws exception otherwise.
-     *
-     * @param ReflectionFunctionAbstract $function
-     * @param ReflectionParameter[]      $parameters
-     * @param mixed[]                    $arguments
-     *
-     * @throws UnknownParameterValueException
-     */
-    private function validateArguments(
-        ReflectionFunctionAbstract $function,
-        array $parameters,
-        array $arguments
-    ) {
-        foreach ($parameters as $num => $parameter) {
-            $name = $parameter->getName();
-
-            if (array_key_exists($num, $arguments) || array_key_exists($name, $arguments)) {
-                continue;
-            }
-
-            throw new UnknownParameterValueException(sprintf(
-                'Can not find a matching value for an argument `$%s` of the method `%s`.',
-                $name,
-                $this->getFunctionPath($function)
-            ));
-        }
-    }
-
-    /**
-     * Returns function path for a provided reflection.
-     *
-     * @param ReflectionFunctionAbstract $function
-     *
-     * @return string
-     */
-    private function getFunctionPath(ReflectionFunctionAbstract $function)
-    {
-        if ($function instanceof ReflectionMethod) {
-            return sprintf(
-                '%s::%s()',
-                $function->getDeclaringClass()->getName(),
-                $function->getName()
-            );
-        }
-
-        return sprintf('%s()', $function->getName());
     }
 
     /**
